added map in station form

+++ database/migrations/2016_12_04_141420_add_coordinates_to_stations_table.php
@@ -13,8 +13,8 @@ class AddCoordinatesToStationsTable extends Migration
     public function up()
     {
         Schema::table('stations', function (Blueprint $table) {
-            $table->decimal('lat')->nullable();
-            $table->decimal('lng')->nullable();
+            $table->double('lat', 30, 26)->nullable();
+            $table->double('lng', 30, 26)->nullable();
         });
     }
